api: make submission concurrency-safe without an explicit transaction
All the intermediary objects before the TestRun can be created freely,
because even if you have concurrent submissions on the same project for
the same build on the same environment, we are already using
get_or_create anyway.

Than the TestRun was being made as 2 steps and that is not necessary. We
now just accumulate the data, and create the object at once at the very
end, so even if there is something wrong with the data and the TestRun
creation fails, it will either create the whole object, or not.