Fix Build name/version migrations
You cannot update a table and modify one of its columns in the same
migration (what implies in the same transaction). We fix this by
splitting the migration in 2: one creates the `name` field, and
initializes it from the version field, and the other changes the type of
the version field.