core: wait a configurable time for test results before notifying
Since test_run results are received asynchronously, we need to assume a
given time window during which a build exists but it's not "finished"
yet, meaning not all test results for it have already arrived. That
window is assumed to be 2 hours by default, so we only notify builds
that are older than 2 hours.

This amount of time can be configured per project, by setting the "build
completion threshold" field in the admin interface.