ci: mark proper TestRun status on awkward LAVA failure
When LAVA reports TestJob as Complete but doesn't produce any result
this might be a result of LAVA bug. It might also be a legitimate
result, but LAVA CI backend of SQUAD doesn't support it. This patch sets
proper flags on TestJob and TestRun objects so the TestJob can be
resubmitted from UI.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>