core/notifications: retry if ProjectStatus id does not exist (yet)
In some cases there could be a timing issue with notifications: a
ProjectStatus is created and a notification is triggered. The
notification request is handled by a worker process before the process
that requested the notification was able to commit the changes to the
database, and the new object is not yet visible to the worker process.

This assumes that id is autoincremented, otherwise a request for a
really unexisting ProjectStatus object would stay in a infinite loop.