core: fix ProjectStatus data
This fixes the pass/fail/skip test count in ProjectStatus, displayed
in the build list, to not count a test duplicated accross different test
runs twice.

This means that the total pass/fail/skip count will sometimes *not*
match the sum of the same values for each test run, where all tests are
counted.