ci/listen: spawn listener processes without fork()
fork()ed subprocess will share the database connection from the master
process, and this will cause errors due to having multiple processes
handling the same connection. The usual sollution for this is closing
the connection before fork()ing, but in this case the master process
also needs its own connection for even starting the subprocesses and
managing their lifecycle. So, we now spawn entirely new processes for
each listener.

This has the disadvantage of making the startup a little slower and
increasing memory consumption. However, we don't have much of a choice
unless we want to completely redesign how listeners work (e.g. by
restricting them from using the database), which is not viable at all at
the moment.