ci: lava: fail gracefully when LAVA replies with error
LAVA sometimes replies with XML-RPC error. In this case there is no
trace that something went wrong. The only indication is that some
results are missing in the frontend. This patch addresses the issue by
failing gracefully, saving error log to the TestJob.failure and sending
admin email notification.

Fixes #97

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>