ci: lava: introduce artificial delay in fetching results
In order to work around the (possible) LAVA issue with saving results,
artificial delay in LAVA backend is introduced. It should give LAVA
enough time to save results in case there are plenty of them. This patch
should be reverted once LAVA problem is fixed or other source of this
issue is identified

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>