ci: Reimplement poll as a series of fetch operations
The poll() task was implemented in a way that Backend.poll() would
directly call fetch() on each test job that needs to be polled. However,
poll() has no exception handling, and therefore an issue with one of the
jobs would cause the entire poll operation to be aborted.

The fetch() task, on the other hand, already handles exceptions, and
does the right thing depending on whether the issue is a temporary or a
permanent one.

Convert Backend.poll() into a generator function that will yield all the
test jobs that should be fetched (instead of fetching them directly).
The actual fetching is then triggered in the background by the poll
task, so that each test job will have exceptions handled individually,
and an issue with one test job will not affect the rest of the test jobs
that are being polled.