core/notification: don't spam admins
If preview notifications don't get marked as already notified, admins
will receive one email per test job after the project timeout has
passed. We only want the admins to be notified once.

When a moderated notification is approved, the final notitification is
sent unconditionally, so it's OK to mark the project status as notified
as sent when sending the preview.