core: associate suites, tests, and metrics with metadata entries
These metadata entries can be edited in the admin interface, and will
serve as a data dictionary of sorts.

Note that while Suite instances are project specific and Test/Metric
instances are created for each specific execution, these metadata
entries are unique; we assume that suites with the same slug in
different projects refer to the same suite in the real world.