submit-ci-jobs: submit ci jobs to a local squad instance
This can be used to locally test the ci integration support using the
fake ci backend, e.g.

./scripts/testdata/gen-ci-jobs myteam foobar

Will create a project myteam/foobar, then submit (by default) 10 fake ci
jobs to it, one per build.

For this to work there must be a squad instance running locally, and
listening to port 8000. `./manage.py runserver` is the easiest way to
accomplish that.