ci/lava: parse suite versions from special metadata keys
LAVA at the moment does not support nested metadata, so we need to
workaround that in order to embed suite version information in LAVA job
definitions.

We have been using ${suite}__version for a while in Linaro, so it is
reused here to populate the `suite_versions` metadata key that SQUAD
expects.