frontend: Add pagination to test and metric listings
For this, the test results and metrics had to be moved away from the
test run page so that we can control which list is being paginated. The
test run page no longer lists the test results and metrics themselves,
but contains links to each individual listing, similar to the build
page.