frontend: paginate "all test results" page
The hiding of failures by default was removed. But tests are sorted by
number of passes, so tests with more failues will still be listed
first, so if there are test failures, they will be listed at page 1.

The filtering was also removed. Since the filtering is client-side, only
the current page of data is available, so filtering would be confusing.