ci: lava: parse LAVA log using incremental parser
LAVA presents the log as YAML document. We're only interested in the
'device' level logs. So all other logs are discarded. In other to do
this operation, YAML content has to be parsed. Loading the whole log
into memory caused OoM issues in the past. This patch uses incremental
parsing approach in order to avoid loading full yaml to memory.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>