frontend: improve display of descriptions in project list
Main design goal: make sure all rows have close to the same height by
default, at least on larger screens. On smaller screens the content will
start to wrap and some rows will be higher than others.

Clicking the "»" on each project will expand the full description, and
then clicking the "«" will collapse it back.

The interaction javascript was greatly simplified, and I realized that
using an Angular controller for that was overkill, so I replaced it
with a few jQuery calls.