ci: lava: update listener to match new LAVA state machine
Due to changes in LAVA state machine, the ZMQ events contain 2 fields
representing jobs status: 'state' and 'health'. 'state' is responsible
for marking the job is running or not. 'health' represents the status of
finished job (Complete, Incomplete, Canceled). This patch adjusts the
logic in lava listener code to work with the new dictionary keys.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>