api: add search and sorting features
This commit adds options to search and order objects in REST API.
Searching is done using "?search" keyword in the URL and uses all fields
specified in the ViewSet. It isn't possible to choose which field to
search. Sorting is done using "?ordering" keyword in URL. Sorting can be
done using one or more fields in ascending or descending order.
Descending order can be achieved by adding "-" in front of the field
name.

Both features only work on simple model fields. Searching or ordering
using related fields (ForeignKey) is not possible at the moment.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>