core: fetch project and group for Test
On my tests, the response time of a single request to the test history
page is cut in half.

Test.prefetch_related is used to load the associated objects for a list
of tests, what happens for example in the "test history" page. Not
preloading project and group data causes a N+1 query situation that
makes the test history page brutally slow.