api: make API web frontend more responsive
With added filtering default API web frontend provides HTML witdgets
which take a lot of time to render. This is best visible with <select>
tag populated with big number of options. In order to mitigate this
problem, widgets were replaced with simple text input. Given that web UI
is not the main use case for the API, this change should not cause any
issues.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>