frontend: allow users to subscribe/unsubscribe to email notifications in a project
Add new panel to user settings which utilizes select2 js lib to allow
user to subscribe to projects in an easy manner.
Display subscriptions and allow for removal of those.
Allow admin to control user subscriptions.
Validate that no email and user can be present at the same time in a
subscription (and that at lease one must be present).