core: cache Status.has_metrics
Instead of making an explicit query every time we just need to know
whether a given test run has metrics for a given suite, we cache that
information when creating the Status object. This improves a lot the
response time of build page with a large-ish number of test runs.

Note: migration 0082_populate_status_has_metrics is skipped on SQLite3
because it does not support "UPDATE ... FROM", what makes the update *a
lot* faster. Nobody should be running production systems with SQLite3
anyway, and this migration is intended to update existing data so we
don't care about SQLite3 in this case.