api: add regressions and fixes to ProjectStatus
In order to present regressions and fixes in the UI they are now cached
in ProjectStatus object for each build. This patch also adds them to
REST API. There is no migration to calculate regressions and fixes for
historical builds as this takes a very long time. Instead a separate
script will be provided.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>