frontend: build: improve column widths in test results table
With only one, or a few environment, the "suite" column would extend too
wide. Invert the logic, fixating a reasonable width for the suite
column, and let the browser best use the width available to the
environment columns.

Also, set a minimum width for the test details table so that it doesn't
get too compressed.