frontend: fix "All test results" with xfail results
(cherry picked from commit 5b310163d8920079b15a66f642503b8fcf5ea0d0)

The change to the tests was reverted because it breaks the build on
travis. There are changes needed to properly sorting of the "All test
results" page in the presence of tests with status xfail, and those are
not suitable for a stable release.

These changes have been have manually verified to fix the crash.