core: NotificationDelivery: allow new notifications on changes
When the notification contents changes, an IntegrityError would be
raised due to the status_id column having a UNIQUE constraint. Changing
the status field from OneToOneField to ForeignKey removes that
constraint, and we end with the only UNIQUE constraint that we should
have, which is the one on all of the fields.

Fixes #363