core: add tracking to EmailTemplate changes
It uses https://github.com/treyhunner/django-simple-history.
Simply added a 'history' field in the EmailTemplate model and set up
some settings to get it integrated with Django Admin.

Closes #169

NOTE1: django-simple-history creates a 1-to-many relationship against the model,
keeping a whole copy of the changed model. There are ways to exclude fields
from being tracked, if needed.

NOTE2: two migrations were needed by this feature: one to create
the HistoricalEmailTemplate model and another one to popuplate it
with current EmailTemplate data.