core, frontend: implement data retention policy
Each project can specify their own data retention policy.
Project.data_retention_days is the numbers of days after which a build
can be deleted. Setting it to 0 (the default) or any negative number
disables the cleanup of old builds completely.

If Project.data_retention_days is set to some positive number, builds
that have the `keep_data` flag set will not be deleted.

Requests to builds that have been cleanup up still return 404, but with
a custom message that let's users know that the build in question has
been removed. Every 404 will now entail an extra database query, but one
that should be pretty fast.