test: make non-stdlib 'mock' package optional
In Python 3.7 at least, unittest.mock has all the features that the
external mock package has. This way, when we move to a newer python as
the standard platform, we will be able to drop this extra test
dependency

mock is currently the the only package outside of the actual
dependencies (requirements.txt) that is needed to run the application
tests.