frontend: support groups and project with empty name
The field name is optional for both group and project created through
the API or via the command line. But groups and projects with empty
names are displayed as `None` throughout the UI.

This patches changes all instances of `group.name` and `project.name` to
use a new `display_name` property, which will fallback to the slug in
case the name is empty.