Add build settings tab
Add build settings as a new tab in `build-nav.jinja2`, thus
allowing end-user to mark whether or not that build should
keep its data in an event of Data Policy Rentention removal.

This will avoid the need of going in admin UI to do the same action,
also usefull for future build-specific settings.