frontend: metricThreshold: fix deleting items
On Firefox, clicking on the trash can icon does delete the threshold in
the database via the API, but there is no visual feedback: the item is
not removed from the list. Looking at the Javascript console on Firefox,
it says:

    ReferenceError: "assignment to undeclared variable index"

Just declaring the variable at the point of assignment fixes it.