frontend: mark strings that compose the home page as translatable
This can be used as an example of how to do it for other parts of the
system.

To create a translation for a module, copy
squad/MODULE/locale/django.pot into
squad/MODULE/locale/LC_MESSAGES/django.po and translate the messages in
it.

To update the .pot files (message catalog template) and merge them into
previously existing translations, run ./scripts/update-translation-files