Dockerfile: cache installed dependencies
Having the installation of packages after WORKDIR and COPY forces a
reinstallation of the dependencies every time there is any change to the source
tree. By installing the dependencies directly on top of the base image, that
layer will be used from the cache and only needs to be rebuilt if the list of
deopendencies change.

The resulting image has a couple of extra layers, but is the same size as
before:

squad/after   latest   3e6bc1fd3f33  6 seconds ago   280MB
squad/before  latest   e09e3b1ead22  2 minutes ago   280MB

(Also, install iproute2 together with auto-apt-proxy so that the later works properly)