api: add subscribe/unsubscribe endpoints to Project
This patch adds /api/project/<id>/subscribe/ and
/api/project/<id>/unsubscribe/ endpoints. They require POST requests
with single field "email". If the user with requested email exists in
the database it's subscribed to all project notifications. If there is
no user with this email, only email address is subscribed. Email address
has to be valid. Unsubscribe works in a similar way by checking email
address of all subscribed users or matchin email only subscriptions. All
of these are removed.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>