squad/controllers/charts.js: display tooltip on line hover
In current chart is quite tricky to get a tooltip to show up.
You need to hover on the specific point to get it displayed, and
that's just a few pixels wide.

With this configuration set up, a tooltip will be displayed
pointing to the nearest point that the mouse is hovering on.

Also, if two points overlap each other, the tooltip will
include both of their data.