frontend: allow empty searches in test comparison page
So far the minimum input required in search boxes in test comparison
page was 1 character. This led to situation where it was hard to use the
form without knowning at least 1st character of the suite or test name.
This patch removes the 1 character limitation and allows for empty
searches.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>