ci: add support for LAVA multinode jobs
This patch adds support for multinode jobs. In case of multnode LAVA API
calls return list in the following form: [job_id.0, job_id.1, ...].
SQUAD will use each of the IDs in the list to create new TestJob object.
This way we ensure that results from all test jobs are assigned to
proper build and fetched.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>