ci: make sure null backend is never used
null CI backend should only be used as a reference for implementing real
ones and in some cases in tests. In some tests null backend methods have
to be properly mocked to avoid wrong results. For this reason all null
backend methods that require to be implemented now raise
NonImplementedError.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>