babel.cfg: add django template extractor
Add django template extractor so that django templates
can also be translated. There is a django-babel package
that does just that, but it has not been updated in a while,
thus lacking support to Django 2.x.

`extract.py` was copied from django-babel to squad/frontend/extract.py
and tweaked lightly in order to provide Django 2.x support.