core: models: improved performance of TestSummary
TestSummary was pre-fetching all tests from all testruns,
which takes a huge slab of the memory as a tradeof of running
less queries.

This patch reduces the amount of memory from hundreds of MB to just
a couple with no difference in run time and still using a couple of queries.
It does a DISTINCT in the database instead of in-memory.