squad: celery: make tasks groups unique on SQS
If task-A has an ETA of 10 minutes gets picked by a worker, a
given task-B with an ETA of 1 minute wouldn't be delivered to
any worker because of the FIFO strategy. Only after 10 minutes,
task-B would've been delivered to workers.

To prevent that behavior, the SQS documentation [1] says that the
`MessageGroupId` should be set to different values. This attribute
is used in case where tasks order are crucial, but that's not the case
of SQUAD when processing notifications.

ref:
[1] https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/using-messagegroupid-property.html