core: models: add environment to MetricThreshold
Add environment to MetricThreshold to replace project as FK. This is the
first commit and individual pr of a set of 3. This part just adds the
environment to metricthreshold model. Client code that uses this new
fields and migrating the data and dropping the project will happen in
the subsequent commits/PRs.