core: tasks: notification: keep only one task for notification on timeout
Keeps only one task for notification timeout.

Before, every new testrun for a given build would call up the task
`maybe_notify_project_status` and therefore a new task for
`notification_timeout` would be added to the queue.

Problem is that if a project has hundres of testruns, this means
hundreds of the exact same call to `notification_timeout`. Because
that task forces sending notification, it'll invalidade all subsequent
attempts to send the same notification.

This patch limits the number of `notification_timeout` to a single one.