ci: models: merge really_fetch within fetch in Backend
This patch prepares Backend code to query a TestJob fetch and flag
it as fetched in a smaller block of code. This is necessary to make
race condition checks when two or more tasks try to fetch the same
TestJob.

Lots of tests have been updated to comply with this refactoring.
"Backend.fetch()" now expects a TestJob id, instead of the object.