frontent: build: include test jobs with no status in Test jobs tab
Users complained that the "Test jobs" tab in the build view showed zero
jobs, but when clicked, it showed a list of jobs.

This happened due to the way that tab was being counted. It didn't count
jobs which "job_status=null".

This patch changes it so that it includes all test jobs despite of their status