core: models: deprecate TestRun and Attachment storage database fields
This patch is part of a series of changes in SQUAD that
moves TestRun and Attachment files away from database into
an external storage.

This part aims to make SQUAD use external files instead of
the ones in the db. It will stay like it for some time until
we're sure enough that everything is working.

The last step will be removing 'old_' fields all alone
and rely on data from files only.