core: comparison: use raw sql to get regressions and fixes
This patch makes it quicker and less memory intensive gettings
regressions and fixes. Currently all tests from both builds are
loaded into memory to just then find regressions and fixes.

The raw SQL is ugly and it should be translated to Django's ORM
if it becomes available. A raw SQL will get all regressions and
fixes in a single trip to DB and will generate more queries if
there are fixed tests with intermittent known issues