core, commands: 4th attempt to make faster fix
Charles changed the default ordering which was making things so slow to
use blank ordering and added multithreads.
A pattern has been identified for all buggy suitemetadata introduced
by the tradefed bug. The suite start with either 'armeabi*' or 'arm64*'.
Using this the command has been rewritten to execute faster. This should
be the best we have in hand right now.