socialaccount: add Github and Gitlab social account login
This will enable users log in to SQUAD by using Gitlab and Github
accounts.

It's important to note that the created user in SQUAD will use the
account's email as username. Therefore, if a user tries to
login with both of his/her Gitlab/Github account and they
both have the same email, so the user will be logged in to
the same account.

The Django user will be shared among those two social account
users.