Dockerfile: use debian buster-backports as base image
This will allow installing Django 2.2, which is the current
LTS and will support all of our dependencies.

This change was mainly required when adding django-allauth to allow
Github/Gitlab login. Its buster version is 0.38, which contains a bug.
The wanted version is 0.44 which supports Django >= 2.2. So I opted
for using Debian's django pre-build package and installed django-allauth
via pip.