test_pending_migrations: temporarily disable this test
This tests fails because SQUAD no longer uses the name column
for Metric and Django detects that it needs removal.

Due to how SQUAD can be deployed and multiple instances can be
running at the same time, we want to allow this migration to be pending
until we get all running instances with the updated code.

Once the name column actually gets removed, I'll re-enable this test.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>