core: comparison: make lighter/faster MetricComparison
MetricComparison now counts with regressions and fixes only, just like
TestComparison does.

It uses MetricThreshold instances that have value=None to determine if
a matching metric (matches threshold name pattern) have regressed or
improved if compared to a previous/baseline build.

For example: if a threshold determines that all metrics named "my-metric"
have is_higher_better=False, this means that if the metric in the current build
is higher than the baseline, the metric will regress, or improve if otherwise.

If a threshold has an explicit value set up, it WILL NOT used to find
regressions or improvements(fixes).

This comparison is done mostly in the database, and a final check is done
in memory against existing thresholds, which is way faster than loading
everything in memory at once.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>