frontend: testjobs_progress: make proportional progress bars
It was requested that the project bars should be proportional to the
number of test jobs each environment has. This way, I introduced a
variable called shrink_factor that determines how a testjob
progress bar should shrink based on its number of testjobs when compared
with the environment with the most jobs.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>