core: templatetags: squad_notification: handle list of lists
If the value received is a list, squad will try to split it in
a nice way. If it's a list of lists, squad would get lost and
crash. Instead of figuring out the depth of the list of lists,
just print out the pythonic representation of the metadata value.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>