ci: models: avoid polling jobs waiting more than a week on the backend
SQUAD has no way to tell whether a TestJob has been worked on by
its backend. It might be that the device is out or that the backend
is undergoing a unusually long maintenance. Overtime, jobs in this
scenario will start clogging up the fetch queue, delaying other fetched
jobs.

I hardcoded this feature for a week, because that's the usual behavior I
noticed, but this can be done via a backend setting as well if requested.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>