ci: Avoid redirection when cancelling or resubmitting LAVA jobs
In standard LAVA deployment calls to REST API that don't end the URL
with slash will be redirected (HTTP 301) to URL with trailing slash.
requests library changes the POST request to GET request in such case
and resulting call is invalid. Since LAVA deals properly with trailing
slashes, the easiest fix is to use them in calls from SQUAD.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@foundries.io>