ci: models: handle resubmission exceptions
When a job runs into a submission issue, it can later be
resubmitted by the user. The resubmission was not catching
any exceptions and the user would not know about the new
exceptions. With this patch, the user will know of
any new issues the test job submission might be running into.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>