ci: models: really re-submit unsubmitted jobs
This sounds weird, but it's simple to explain: occasionaly jobs
fail to get submitted, therefore resubmitting them would require
a job_id. That's how LAVA resubmits jobs.

But in SQUAD, when trying to resubmit a job without job_id, the same
procedure of submit will take place, therefore needing a few changes
in tests where no job_id was specified and "resubmit" was expected
to be executed.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>