frontend: templates: testjobs: avoid query for perent jobs
When invoking `testjob.resubmitted_jobs.count()` it triggers a
query for testjobs looking up the parent_job_id column.

Later this count is used only to determine if a button show be
displayed or not. So I just replaced the extra query by the
pre-calculated testjob.resubmitted_count, which contains exactly
the same information already.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>