frontend: failures: compute a confidence score
It is difficult to tell if a test failure is a true regression or a
flakey test. To try and address this, we create a new view of test
failures for a build that displays a confidence score.

A confidence score is the ratio of the number of times the test has
passed over some number number of times the test has been run.

Signed-off-by: Justin Cook <justin.cook@linaro.org>