ci: tuxsuite: handle failed jobs
The TuxSuite backend was ignoring all test jobs that had
failed results. There can be two reasons why a test job fails:
a) the build failed or b) the tests in the job actually failed.

SQUAD will now handle the second scenario as a regular job, except
that the failure will be saved the testjob.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>