frontend: build: add redirect URL to get API URL
When data is submitted via the API, there is currently no structured way
to obtain the internal build ID so that it can be used for API requests.
Add an URL at /<project>/<build>/build/<version>/api/ that will redirect
to the correct API endpoint for the given build, i.e. /api/builds/<id>/.