frontend/templatetags: Add handling for missing status
If status was missing, this would previously cause a 500 error when
loading the page, due to a RelatedObjectDoesNotExist error being raised
when the project status function was called.

Now, add handling for the case where status is None, so the
project_status will return None instead of raising an error.

Signed-off-by: Katie Worton <katie.worton@linaro.org>