core/models,frontend/user_settings: Add user preference for failures only
Add support for a user preference for the "failures only" field in the
build results screen.
When the user is logged in, they can set a user preference in their
profile for this "failures only" field. The default for this field is
for the "failures only" tickbox to be checked.

Signed-off-by: Katie Worton <katie.worton@linaro.org>