ci/tasks.py: add new task to process plugins with subtasks
Commit 47a5e52a96ba776ba9aee8ab26debce917202273 split plugin
post-processing into a different task, to alleviate the load
on the main task. Unfortunatelly it had a side effect that
made all post-processing tasks to be consumed only by one worker
causing a general slowdown fetching results.

This new patch still uses the same idea, except that only tasks
for plugins with subtasks will be consumed by this one worker.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>