core/models.py: prevend malformed yaml to be saved
ProjectStatus stores build statistics including regressions and fixes. These
are stored in serialized yaml objects which should have been validated before
save. This was not the case.

This patch adds a validation on those fields before every save and also adds code
that prevents a crash if SQUAD tries to access a fields with bad previous values.

From now on, no malformed yaml will be accepted.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>